# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "modo": {
            "default": "propostas",
            "title": "Modo",
            "type": "string"
          },
          "pergunta": {
            "title": "Pergunta",
            "type": "string"
          }
        },
        "required": [
          "pergunta"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ConfigItem": {
        "properties": {
          "categoria": {
            "title": "Categoria",
            "type": "string"
          },
          "chave": {
            "title": "Chave",
            "type": "string"
          },
          "configurado": {
            "title": "Configurado",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "origem": {
            "title": "Origem",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "secreto": {
            "title": "Secreto",
            "type": "boolean"
          },
          "valor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor"
          }
        },
        "required": [
          "chave",
          "label",
          "categoria",
          "secreto",
          "configurado",
          "origem"
        ],
        "title": "ConfigItem",
        "type": "object"
      },
      "ConfigSet": {
        "properties": {
          "chave": {
            "title": "Chave",
            "type": "string"
          },
          "valor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor"
          }
        },
        "required": [
          "chave"
        ],
        "title": "ConfigSet",
        "type": "object"
      },
      "ConformidadeRead": {
        "properties": {
          "cache_atualizado_em": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Atualizado Em"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "detalhe": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalhe"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "municipio_ibge": {
            "title": "Municipio Ibge",
            "type": "string"
          },
          "numero": {
            "title": "Numero",
            "type": "string"
          },
          "orgao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orgao"
          },
          "secao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secao"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tipo": {
            "title": "Tipo",
            "type": "string"
          },
          "validade": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validade"
          },
          "valor": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor"
          }
        },
        "required": [
          "id",
          "municipio_ibge",
          "tipo",
          "numero"
        ],
        "title": "ConformidadeRead",
        "type": "object"
      },
      "ConformidadeResumo": {
        "description": "CAUC: KPIs por status + por seção. CAPAG: rating separado.",
        "properties": {
          "a_comprovar": {
            "title": "A Comprovar",
            "type": "integer"
          },
          "capag": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConformidadeRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "comprovados": {
            "title": "Comprovados",
            "type": "integer"
          },
          "desativados": {
            "title": "Desativados",
            "type": "integer"
          },
          "requisitos": {
            "items": {
              "$ref": "#/components/schemas/ConformidadeRead"
            },
            "title": "Requisitos",
            "type": "array"
          },
          "secoes": {
            "items": {
              "$ref": "#/components/schemas/SecaoResumo"
            },
            "title": "Secoes",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "comprovados",
          "a_comprovar",
          "desativados",
          "secoes",
          "requisitos"
        ],
        "title": "ConformidadeResumo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IslandRequest": {
        "properties": {
          "pergunta": {
            "title": "Pergunta",
            "type": "string"
          }
        },
        "required": [
          "pergunta"
        ],
        "title": "IslandRequest",
        "type": "object"
      },
...
      "LlmChaveIn": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "LlmChaveIn",
        "type": "object"
      },
      "LlmModelosOut": {
        "properties": {
          "erro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erro"
          },
          "modelos": {
            "items": {
              "type": "string"
            },
            "title": "Modelos",
            "type": "array"
          },
          "origem": {
            "title": "Origem",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "modelos",
          "origem"
        ],
        "title": "LlmModelosOut",
        "type": "object"
      },
      "LlmProviderOut": {
        "properties": {
          "chave": {
            "title": "Chave",
            "type": "string"
          },
          "chave_mascarada": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chave Mascarada"
          },
          "configurado": {
            "title": "Configurado",
            "type": "boolean"
          },
          "docs_url": {
            "title": "Docs Url",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "modelo_chat": {
            "default": false,
            "title": "Modelo Chat",
            "type": "boolean"
          },
          "modelo_resumo": {
            "default": false,
            "title": "Modelo Resumo",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "label",
          "chave",
          "configurado",
          "docs_url"
        ],
        "title": "LlmProviderOut",
        "type": "object"
      },
...
      "SecaoResumo": {
        "properties": {
          "a_comprovar": {
            "title": "A Comprovar",
            "type": "integer"
          },
          "comprovados": {
            "title": "Comprovados",
            "type": "integer"
          },
          "desativados": {
            "title": "Desativados",
            "type": "integer"
          },
          "secao": {
            "title": "Secao",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "secao",
          "total",
          "comprovados",
          "a_comprovar",
          "desativados"
        ],
        "title": "SecaoResumo",
        "type": "object"
      },
...
      "SyncConformidadeRequest": {
        "properties": {
          "municipio_ibge": {
            "maxLength": 7,
            "minLength": 7,
            "title": "Municipio Ibge",
            "type": "string"
          }
        },
        "required": [
          "municipio_ibge"
        ],
        "title": "SyncConformidadeRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "API pública v1 — propostas/repasses do governo (cache-first, RLS por usuário).",
    "title": "Hub Capture API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/config": {
      "get": {
        "operationId": "listar_config_api_v1_admin_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConfigItem"
                  },
                  "title": "Response Listar Config Api V1 Admin Config Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Listar Config",
        "tags": [
          "admin-config"
        ]
      },
      "put": {
        "operationId": "definir_config_api_v1_admin_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConfigItem"
                  },
                  "title": "Response Definir Config Api V1 Admin Config Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Definir Config",
        "tags": [
          "admin-config"
        ]
      }
    },
    "/api/v1/admin/config/llm/providers": {
      "get": {
        "description": "Provedores de LLM suportados, com status da chave e seleção de modelo atual.",
        "operationId": "listar_llm_providers_api_v1_admin_config_llm_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LlmProviderOut"
                  },
                  "title": "Response Listar Llm Providers Api V1 Admin Config Llm Providers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Listar Llm Providers",
        "tags": [
          "admin-config"
        ]
      }
    },
    "/api/v1/admin/config/llm/{provider_id}/chave": {
      "put": {
        "description": "Grava a API key do provedor e já devolve os modelos dele (menor fricção).",
        "operationId": "definir_llm_chave_api_v1_admin_config_llm__provider_id__chave_put",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmChaveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmModelosOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Definir Llm Chave",
        "tags": [
          "admin-config"
        ]
      }
    },
    "/api/v1/admin/config/llm/{provider_id}/modelos": {
      "get": {
        "description": "Modelos do provedor usando a chave já salva (listagem ao vivo + fallback).",
        "operationId": "listar_llm_modelos_api_v1_admin_config_llm__provider_id__modelos_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmModelosOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Listar Llm Modelos",
        "tags": [
          "admin-config"
        ]
      }
    },
...
    "/api/v1/compliance": {
      "get": {
        "operationId": "conformidade_resumo_api_v1_compliance_get",
        "parameters": [
          {
            "description": "código IBGE (7 dígitos)",
            "in": "query",
            "name": "municipio",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "código IBGE (7 dígitos)",
              "title": "Municipio"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConformidadeResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Conformidade Resumo",
        "tags": [
          "conformidade"
        ]
      }
    },
    "/api/v1/compliance/sync": {
      "post": {
        "operationId": "conformidade_sync_api_v1_compliance_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncConformidadeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Conformidade Sync Api V1 Compliance Sync Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Conformidade Sync",
        "tags": [
          "conformidade"
        ]
      }
    },
    "/api/v1/copilot/chat": {
      "post": {
        "operationId": "copiloto_chat_api_v1_copilot_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Copiloto Chat",
        "tags": [
          "copiloto"
        ]
      }
    },
    "/api/v1/copilot/island": {
      "post": {
        "description": "Agente do Dynamic Island (tool calling). O loop de ferramentas roda AQUI,\ncom a sessão RLS do request viva; o stream só replaye os eventos coletados\n(mesma razão do RAG pré-stream acima).",
        "operationId": "copiloto_island_api_v1_copilot_island_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IslandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Copiloto Island",
        "tags": [
          "copiloto"
        ]
      }
    },
...
  }
}